Validation to maximum number of salvoes added to backend
If owner is the creator, player can fire as many location as ship number left on the last turn.
If owner is joined, can wait to the next turn because the creator has preferency of turn.
If creator sinks the joiners ships in a turn then joiner, lost some salvo location rights even before playing its own turn.
This same is applied to frontend too.